+++ src/Modules/Enforcement/interfaces/EnforcementCriteriaInterface.sol
@@ -2,48 +2,41 @@
 
 pragma solidity ^0.8.17;
 
+import { EnumerableSet } from '@openzeppelin/contracts/utils/structs/EnumerableSet.sol';
+
 interface EnforcementCriteriaInterface {
+    /// @dev The relevant scoring criteria for a request.
+    struct Buckets {
+        uint256[] ranges;
+        uint256[] weights;
+    }
 
-    /**
-     * @notice Allows a maintainer to review and updates the payout state.
-     * @param _submissionId The submission to review.
-     * @param _score The score to give the submission.
-     */
-    function review(
-          uint256 _submissionId
-        , uint256 _score
-    )
-        external;
+    /// @dev The scores given to a service submission.
+    struct Score {
+        uint256 reviewSum;
+        uint256 scaledAvg;
+        EnumerableSet.AddressSet enforcers;
+    }
 
-    /**
-     * @notice Calculates the amount of payment tokens a submission is entitled to.
-     * @param _laborMarket The labor market the submission is in.
-     * @param _submissionId The submission to calculate the payment reward for.
-     * @return The amount of payment tokens a submission is entitled to.
-     */
-    function getReward(
-          address _laborMarket
-        , uint256 _submissionId
-    )
-        external
-        view
-        returns (
-            uint256
-        );
+    /// @dev The relevant storage data for a request.
+    struct Request {
+        uint256 scaledAvgSum;
+        mapping(uint256 => Score) scores;
+        EnumerableSet.AddressSet qualifiedProviders;
+    }
 
-    /** 
-     *  @notice Get the remaining payment tokens not allocated.
-     *  @param _laborMarket The labor market the request is in.
-     *  @param _requestId The request to get the remaining unclaimed for.
-     *  @return unclaimed Total that can be reclaimed by the requester.
-     */
-    function getRemainder(
-          address _laborMarket
-        , uint256 _requestId
-    ) 
-        external 
-        view 
-        returns (
-            uint256
-        );
+    function enforce(
+        uint256 _requestId,
+        uint256 _submissionId,
+        uint256 _score,
+        address _enforcer
+    ) external returns (bool, uint24);
+
+    function rewards(
+        uint256 _requestId,
+        uint256 _submissionId,
+        uint256 _pTokenQ
+    ) external returns (uint256 amount, bool requiresSubmission);
+
+    function remainder(uint256 _requestId, uint256 _pTokenQ) external returns (uint256);
 }